requirements.txt
google-adk>=1.25.0
google-genai>=1.0.0
google-cloud-aiplatform>=1.74.0
pybreaker>=1.4.0
